fixed ranking not being available at every level:   - moved ranking event handling back to base component   - moved ranking functionality to separate service
  Deleted some invalid/unused import statements.
