+++ src/app/cardsView/project/project.component.ts
@@ -4,8 +4,8 @@
 import { Component } from '@angular/core';
 import {ActivatedRoute} from "@angular/router";
 import {ProjectService} from "../services/project/project.service";
-import {ParentCard} from "../../card/card";
 import {BasePaginateCardComponent} from "../base-paginate-card.component";
+import {RankService} from "../services/rank/rank.service";
 
 
 @Component({
@@ -16,8 +16,8 @@ import {BasePaginateCardComponent} from "../base-paginate-card.component";
 export class ProjectComponent extends BasePaginateCardComponent {
   private key: number;
 
-  constructor(private route: ActivatedRoute, private service: ProjectService) {
-    super();
+  constructor(private route: ActivatedRoute, private service: ProjectService, rankService: RankService) {
+    super(rankService);
   }
 
   getPage(page: number) {
