+++ src/app/cardsView/services/base-card-mock.service.ts
@@ -31,8 +31,4 @@ export abstract class BaseCardMockService {
       .delay(delayTime)
       .map(res => <PaginateResponse<ParentCard>>{ total: this.total, cards: res.json().slice(start, start + size)});
   }
-
-  rankIssue(id: string, direction: string, referenceId: string) {
-    console.log(id, direction, referenceId);
-  }
 }
